requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-ads1x15==2.4.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
args==0.1.0
binho-host-adapter==0.1.6
clint==0.5.1
numpy==2.1.2
pydot==3.0.2
pyftdi==0.55.4
pyparsing==3.1.4
pyserial==3.5
python-statemachine==2.3.6
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
